Moved full operations to separate file to have clearer structure.
